Refector: Header, page layout 분리
