+++ src/components/UI/index.ts
@@ -2,4 +2,3 @@ export { default as Spacing } from './Spacing';
 export { default as Container } from './Container';
 export { default as Title } from './Title';
 export { default as Button } from './Button';
-export { default as Header } from './Header';
