+++ src/pages/Auth/SignUpPage.tsx
@@ -1,4 +1,3 @@
-// @ts-nocheck
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import useValidation from './hooks/useValidation';
