+++ src/pages/Auth/hooks/useValidation.ts
@@ -1,5 +1,4 @@
 import { useState, useEffect } from 'react';
-
 interface FormProps {
   email: string;
   password: string;
